pyproject.toml
[project]
name = "acme-dental-ai-agent"
version = "0.0.1"
description = "AI Agent for the Acme Dental Clinic"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain[anthropic]>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "pinecone-client>=3.0.0",
    "langchain-community>=0.0.10",
    "langchain-pinecone>=0.1.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

